add flag to save ply and move save_ply to utils
